Refactored streamlit_app.py
