+++ components/__init__.py
@@ -1,0 +1,1 @@
+# This file is intentionally left empty to make the components directory a Python package
